+++ apps/client/app/layout.tsx
@@ -1,8 +1,8 @@
-import './globals.css';
-import StyledJsxRegistry from './registry';
+import '@/app/globals.css';
+import StyledJsxRegistry from '@/app/registry';
 import MyHeader from '@/components/base/MyHeader';
 import MyFooter from '@/components/base/MyFooter';
-import GlobalStateRoot from './GlobalStateRoot';
+import GlobalStateRoot from '@/app/GlobalStateRoot';
 
 export default function RootLayout({ children }: { children: React.ReactNode }): JSX.Element {
   return (
