+++ apps/client/components/ui/wrap/Wrap.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 import { MEDIAQUERY } from '@/constants/constant';
 
-import * as Types from '../types/wrap';
+import * as Types from '@/components/ui/types/wrap';
 
 export const Wrap_mediaquery = styled.div<Types.WrapMediaqueryProp>`
   background-color: white;
