+++ apps/client/containers/main/index.tsx
@@ -3,8 +3,8 @@
 import styled from 'styled-components';
 
 import { FONT, CONST_MAIN_PAGE, MEDIAQUERY } from '@/constants/constant';
+import * as mainTypes from '@/containers/types/main';
 
-import * as mainTypes from '../types/main';
 import { TitleAndText, TitleAndTestBig, TitleAndTestsSmallForSeveral } from '@/components/base/TestContent';
 import { YellowButton } from '@/components/ui/button/Button';
 import { Wrap_mediaquery } from '@/components/ui/wrap/Wrap';
