+++ apps/client/public/images/footer/index.ts
@@ -1,4 +1,4 @@
-import GitHubImage from './gitHub.svg';
-import InstagramImage from './instagram.png';
+import GitHubImage from '@/public/images/footer/gitHub.svg';
+import InstagramImage from '@/public/images/footer/instagram.png';
 
 export { GitHubImage, InstagramImage };
