+++ apps/client/public/images/test/index.ts
@@ -1,3 +1,3 @@
-import TestPlayCountImage from './playCount.svg';
+import TestPlayCountImage from '@/public/images/test/playCount.svg';
 
 export { TestPlayCountImage };
